server: main_router: merge in complete main_controller
It does not make sense to keep the router and controller with their
concerns separated this way. When everything is back in the router, a new
sequence of refactorings can be started.
